# meteo_rest_fastapi/meteo_rest_fastapi/openapi(1).json
{"openapi":"3.1.0","info":{"title":"MeteoApp REST API","description":"Transformación del sistema de monitorización meteorológica basado en sockets/RMI a una API REST con HTTP y JSON. La API mantiene las operaciones LIST, GET, SUB y UNSUB como recursos REST.","contact":{"name":"Grupo MeteoApp"},"license":{"name":"Uso académico"},"version":"1.0.0"},"paths":{"/health":{"get":{"tags":["Sistema"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/cities":{"get":{"tags":["Ciudades"],"summary":"List Cities","description":"Equivalente REST del comando LIST / método remoto listCities().","operationId":"list_cities_cities_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CityListResponse"}}}}}}},"/weather":{"get":{"tags":["Meteorología"],"summary":"Get Default Weather","description":"Consulta puntual del clima. Equivale a GET sin ciudad explícita en el protocolo original.","operationId":"get_default_weather_weather_get","parameters":[{"name":"city","in":"query","required":false,"schema":{"type":"string","description":"Ciudad consultada; por defecto Madrid","default":"Madrid","title":"City"},"description":"Ciudad consultada; por defecto Madrid"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/weather/{city}":{"get":{"tags":["Meteorología"],"summary":"Get Weather","description":"Equivalente REST del comando GET / método remoto getWeather(city).","operationId":"get_weather_weather__city__get","parameters":[{"name":"city","in":"path","required":true,"schema":{"type":"string","title":"City"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeatherResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clients":{"post":{"tags":["Clientes"],"summary":"Create Client","description":"Crea una identidad de cliente REST.\nSustituye la conexión TCP o el objeto callback RMI usado para asociar suscripciones.","operationId":"create_client_clients_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreateResponse"}}}}}}},"/clients/{client_id}":{"delete":{"tags":["Clientes"],"summary":"Delete Client","description":"Elimina el cliente y todas sus suscripciones, equivalente a cerrar la conexión.","operationId":"delete_client_clients__client_id__delete","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clients/{client_id}/subscriptions":{"post":{"tags":["Suscripciones"],"summary":"Create Subscription","description":"Equivalente REST del comando SUB / método remoto subscribe(city, callback).","operationId":"create_subscription_clients__client_id__subscriptions_post","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Suscripciones"],"summary":"List Subscriptions","description":"Lista las suscripciones activas de un cliente REST.","operationId":"list_subscriptions_clients__client_id__subscriptions_get","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Suscripciones"],"summary":"Delete All Subscriptions","description":"Equivalente REST de UNSUB sin ciudad: cancela todas las suscripciones del cliente.","operationId":"delete_all_subscriptions_clients__client_id__subscriptions_delete","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clients/{client_id}/subscriptions/{city}":{"delete":{"tags":["Suscripciones"],"summary":"Delete Subscription","description":"Equivalente REST de UNSUB con ciudad: cancela una suscripción concreta.","operationId":"delete_subscription_clients__client_id__subscriptions__city__delete","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}},{"name":"city","in":"path","required":true,"schema":{"type":"string","title":"City"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clients/{client_id}/notifications":{"get":{"tags":["Notificaciones"],"summary":"Get Notifications","description":"Sustituto REST del callback RMI onWeatherChange(data).\nEn lugar de invocar al cliente, el servidor expone las notificaciones como recurso consultable.","operationId":"get_notifications_clients__client_id__notifications_get","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"string","title":"Client Id"}},{"name":"consume","in":"query","required":false,"schema":{"type":"boolean","description":"Si es true, vacía la cola tras devolverla","default":true,"title":"Consume"},"description":"Si es true, vacía la cola tras devolverla"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CityListResponse":{"properties":{"status":{"type":"integer","title":"Status","default":200},"type":{"type":"string","title":"Type","default":"RESP_LIST"},"data":{"items":{"type":"string"},"type":"array","title":"Data"},"msg":{"type":"string","title":"Msg"}},"type":"object","required":["data","msg"],"title":"CityListResponse"},"ClientCreateResponse":{"properties":{"status":{"type":"integer","title":"Status","default":201},"client_id":{"type":"string","title":"Client Id"},"msg":{"type":"string","title":"Msg"}},"type":"object","required":["client_id","msg"],"title":"ClientCreateResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"status":{"type":"integer","title":"Status"},"msg":{"type":"string","title":"Msg"}},"type":"object","required":["status","msg"],"title":"MessageResponse"},"Notification":{"properties":{"type":{"type":"string","title":"Type","default":"NOTIF"},"city":{"type":"string","title":"City"},"data":{"additionalProperties":{"type":"number"},"type":"object","title":"Data"},"timestamp":{"type":"string","title":"Timestamp"}},"type":"object","required":["city","data","timestamp"],"title":"Notification"},"NotificationListResponse":{"properties":{"status":{"type":"integer","title":"Status","default":200},"client_id":{"type":"string","title":"Client Id"},"data":{"items":{"$ref":"#/components/schemas/Notification"},"type":"array","title":"Data"},"msg":{"type":"string","title":"Msg"}},"type":"object","required":["client_id","data","msg"],"title":"NotificationListResponse"},"SubscriptionCreate":{"properties":{"city":{"type":"string","title":"City","description":"Ciudad a monitorizar","default":"Madrid"},"variables":{"items":{"type":"string"},"type":"array","title":"Variables","description":"Variables meteorológicas a monitorizar"}},"type":"object","title":"SubscriptionCreate"},"SubscriptionInfo":{"properties":{"city":{"type":"string","title":"City"},"variables":{"items":{"type":"string"},"type":"array","title":"Variables"},"last_sent":{"$ref":"#/components/schemas/WeatherValues"}},"type":"object","required":["city","variables","last_sent"],"title":"SubscriptionInfo"},"SubscriptionListResponse":{"properties":{"status":{"type":"integer","title":"Status","default":200},"client_id":{"type":"string","title":"Client Id"},"data":{"items":{"$ref":"#/components/schemas/SubscriptionInfo"},"type":"array","title":"Data"}},"type":"object","required":["client_id","data"],"title":"SubscriptionListResponse"},"SubscriptionResponse":{"properties":{"status":{"type":"integer","title":"Status","default":201},"msg":{"type":"string","title":"Msg"},"client_id":{"type":"string","title":"Client Id"},"city":{"type":"string","title":"City"},"variables":{"items":{"type":"string"},"type":"array","title":"Variables"},"current":{"$ref":"#/components/schemas/WeatherValues"}},"type":"object","required":["msg","client_id","city","variables","current"],"title":"SubscriptionResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeatherResponse":{"properties":{"status":{"type":"integer","title":"Status","description":"Código de estado de la operación","default":200},"city":{"type":"string","title":"City"},"data":{"$ref":"#/components/schemas/WeatherValues"}},"type":"object","required":["city","data"],"title":"WeatherResponse"},"WeatherValues":{"properties":{"temp":{"type":"number","title":"Temp","description":"Temperatura en grados Celsius"},"hum":{"type":"number","title":"Hum","description":"Humedad relativa en porcentaje"},"pres":{"type":"number","title":"Pres","description":"Presión atmosférica en hPa"},"wind":{"type":"number","title":"Wind","description":"Velocidad del viento en m/s"}},"type":"object","required":["temp","hum","pres","wind"],"title":"WeatherValues"}}}}